doc(init): Add documents to initiate project
Created a directory for images (assets) and added
downloaded images for the development of the project.
Created a directory for the resources
(the wireframe, typography/color palette, and a
WORKFLOW.md) required for the development of the
project. Created a WORKFLOW.md file that will detail
the development process of this project. Created an
index.html file and a style.css file.